+++ libs/gui/mdichild.h
@@ -538,6 +538,6 @@ private:
 	std::map<size_t, std::shared_ptr<iADataSetRenderer>> m_3dMagicLensRenderers; //!< 3D renderers for magic lens (one per dataset in m_datasets)
 	std::map<size_t, std::shared_ptr<iASliceRenderer>> m_sliceRenderers;  //!< slice renderers (one per dataset in m_datsets)
 
-	void setDataSetMovable(int dataSetIdx);
+	void setDataSetMovable(size_t dataSetIdx);
 	vtkSmartPointer<iAvtkInteractStyleActor> m_manualMoveStyle[4];   // 
 };
